pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "healthai-assistant"
version = "1.0.0"
description = "AWS-native health information assistant powered by Bedrock and OpenSearch"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = [
    "healthcare",
    "chatbot",
    "aws",
    "bedrock",
    "opensearch",
    "langchain",
    "rag"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "flask>=3.0.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-aws>=0.2.0",
    "boto3>=1.34.0",
    "opensearch-py>=2.4.0",
    "requests-aws4auth>=1.2.0",
    "sentence-transformers>=4.0.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/kiran3035/healthai-assistant"
Documentation = "https://github.com/kiran3035/healthai-assistant#readme"
Repository = "https://github.com/kiran3035/healthai-assistant"
Issues = "https://github.com/kiran3035/healthai-assistant/issues"

[project.scripts]
healthai = "run:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "services*", "config*", "web*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
